+++ src/integrationTest/groovy/wooga/gradle/snyk/SnykPluginIntegrationSpec.groovy
@@ -111,15 +111,9 @@ class SnykPluginIntegrationSpec extends SnykIntegrationSpec {
         Test     | "printJson"                      | true                                 | "Boolean"                         | ConventionSource.extension(extensionName, property)              | _          || _
         Report   | "printJson"                      | true                                 | "Boolean"                         | ConventionSource.extension(extensionName, property)              | _          || _
         Monitor  | "printJson"                      | true                                 | "Boolean"                         | ConventionSource.extension(extensionName, property)              | _          || _
-        Test     | "jsonOutputPath"                 | osPath("/path/to/json1")             | "RegularFile"                     | ConventionSource.extension(extensionName, property)              | _          || _
-        Report   | "jsonOutputPath"                 | osPath("/path/to/json2")             | "RegularFile"                     | ConventionSource.extension(extensionName, property)              | _          || _
-        Monitor  | "jsonOutputPath"                 | osPath("/path/to/json3")             | "RegularFile"                     | ConventionSource.extension(extensionName, property)              | _          || _
         Test     | "printSarif"                     | true                                 | "Boolean"                         | ConventionSource.extension(extensionName, property)              | _          || _
         Report   | "printSarif"                     | true                                 | "Boolean"                         | ConventionSource.extension(extensionName, property)              | _          || _
         Monitor  | "printSarif"                     | true                                 | "Boolean"                         | ConventionSource.extension(extensionName, property)              | _          || _
-        Test     | "sarifOutputPath"                | osPath("/path/to/sarif1")            | "RegularFile"                     | ConventionSource.extension(extensionName, property)              | _          || _
-        Report   | "sarifOutputPath"                | osPath("/path/to/sarif2")            | "RegularFile"                     | ConventionSource.extension(extensionName, property)              | _          || _
-        Monitor  | "sarifOutputPath"                | osPath("/path/to/sarif3")            | "RegularFile"                     | ConventionSource.extension(extensionName, property)              | _          || _
         Test     | "severityThreshold"              | SeverityThresholdOption.critical     | "SeverityThresholdOption"         | ConventionSource.extension(extensionName, property)              | _          || _
         Report   | "severityThreshold"              | SeverityThresholdOption.critical     | "SeverityThresholdOption"         | ConventionSource.extension(extensionName, property)              | _          || _
         Monitor  | "severityThreshold"              | SeverityThresholdOption.low          | "SeverityThresholdOption"         | ConventionSource.extension(extensionName, property)              | _          || _
@@ -201,10 +195,9 @@ class SnykPluginIntegrationSpec extends SnykIntegrationSpec {
         Monitor  | "projectBusinessCriticality"     | [BusinessCriticalityOption.critical] | "List<BusinessCriticalityOption>" | ConventionSource.extension(extensionName, property)              | _          || _
         Monitor  | "projectTags"                    | ["foo": "bar"]                       | "Map<String, String>"             | ConventionSource.extension(extensionName, property)              | _          || _
 
-        Test     | "reports.sarif.enabled"          | true                                 | "Boolean"                         | ConventionSource.extension(extensionName, "sarifReportsEnabled") | ""         || _
-        Test     | "reports.json.enabled"           | true                                 | "Boolean"                         | ConventionSource.extension(extensionName, "jsonReportsEnabled")  | ""         || _
-        Report   | "reports.sarif.enabled"          | true                                 | "Boolean"                         | ConventionSource.extension(extensionName, "sarifReportsEnabled") | ""         || _
-        Report   | "reports.json.enabled"           | false                                | "Boolean"                         | ConventionSource.extension(extensionName, "jsonReportsEnabled")  | ""         || _
+        Test     | "reports.sarif.required"         | true                                 | "Boolean"                         | ConventionSource.extension(extensionName, "sarifReportsEnabled") | _          || _
+        Test     | "reports.json.required"          | true                                 | "Boolean"                         | ConventionSource.extension(extensionName, "jsonReportsEnabled")  | _          || _
+        Report   | "reports.sarif.required"         | true                                 | "Boolean"                         | ConventionSource.extension(extensionName, "sarifReportsEnabled") | _          || _
 
         Test     | "debug"                          | true                                 | "Boolean"                         | ConventionSource.extension(extensionName, property)              | _          || _
         Monitor  | "debug"                          | true                                 | "Boolean"                         | ConventionSource.extension(extensionName, property)              | _          || _
@@ -376,14 +369,6 @@ class SnykPluginIntegrationSpec extends SnykIntegrationSpec {
         "printJson"                      | toProviderSet(property) | true                                                            | _                                                                      | "Provider<Boolean>"                         | PropertyLocation.script
         "printJson"                      | toProviderSet(property) | false                                                           | _                                                                      | "Boolean"                                   | PropertyLocation.script
 
-        "jsonOutputPath"                 | _                       | _                                                               | null                                                                   | "Provider<RegularFile>"                     | PropertyLocation.none
-        "jsonOutputPath"                 | _                       | osPath("/path/to/json1")                                        | _                                                                      | _                                           | PropertyLocation.environment
-        "jsonOutputPath"                 | _                       | osPath("/path/to/json2")                                        | _                                                                      | _                                           | PropertyLocation.property
-        "jsonOutputPath"                 | toSetter(property)      | osPath("/path/to/json3")                                        | _                                                                      | "Provider<RegularFile>"                     | PropertyLocation.script
-        "jsonOutputPath"                 | toSetter(property)      | osPath("/path/to/json4")                                        | _                                                                      | "File"                                      | PropertyLocation.script
-        "jsonOutputPath"                 | toProviderSet(property) | osPath("/path/to/json5")                                        | _                                                                      | "Provider<RegularFile>"                     | PropertyLocation.script
-        "jsonOutputPath"                 | toProviderSet(property) | osPath("/path/to/json6")                                        | _                                                                      | "File"                                      | PropertyLocation.script
-
         "printSarif"                     | _                       | _                                                               | false                                                                  | "Provider<Boolean>"                         | PropertyLocation.none
         "printSarif"                     | _                       | "true"                                                          | true                                                                   | _                                           | PropertyLocation.environment
         "printSarif"                     | _                       | "true"                                                          | true                                                                   | _                                           | PropertyLocation.property
@@ -391,14 +376,6 @@ class SnykPluginIntegrationSpec extends SnykIntegrationSpec {
         "printSarif"                     | toSetter(property)      | false                                                           | _                                                                      | "Boolean"                                   | PropertyLocation.script
         "printSarif"                     | toProviderSet(property) | true                                                            | _                                                                      | "Provider<Boolean>"                         | PropertyLocation.script
         "printSarif"                     | toProviderSet(property) | false                                                           | _                                                                      | "Boolean"                                   | PropertyLocation.script
-
-        "sarifOutputPath"                | _                       | _                                                               | null                                                                   | "Provider<RegularFile>"                     | PropertyLocation.none
-        "sarifOutputPath"                | _                       | osPath("/path/to/sarif1")                                       | _                                                                      | _                                           | PropertyLocation.environment
-        "sarifOutputPath"                | _                       | osPath("/path/to/sarif2")                                       | _                                                                      | _                                           | PropertyLocation.property
-        "sarifOutputPath"                | toSetter(property)      | osPath("/path/to/sarif3")                                       | _                                                                      | "Provider<RegularFile>"                     | PropertyLocation.script
-        "sarifOutputPath"                | toSetter(property)      | osPath("/path/to/sarif4")                                       | _                                                                      | "File"                                      | PropertyLocation.script
-        "sarifOutputPath"                | toProviderSet(property) | osPath("/path/to/sarif5")                                       | _                                                                      | "Provider<RegularFile>"                     | PropertyLocation.script
-        "sarifOutputPath"                | toProviderSet(property) | osPath("/path/to/sarif6")                                       | _                                                                      | "File"                                      | PropertyLocation.script
 
         "severityThreshold"              | _                       | _                                                               | null                                                                   | "Provider<SeverityThresholdOption>"         | PropertyLocation.none
         "severityThreshold"              | _                       | SeverityThresholdOption.low                                     | _                                                                      | _                                           | PropertyLocation.environment
