--- src/main/groovy/wooga/gradle/snyk/SnykActionSpec.groovy
@@ -1,33 +1,0 @@
-/*
- * Copyright 2022 Wooga GmbH
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *        http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package wooga.gradle.snyk
-
-
-import org.gradle.api.provider.Provider
-
-interface SnykActionSpec {
-    Provider<List<String>> getArguments()
-
-    void setAdditionalArguments(Iterable<String> value)
-    void setAdditionalArguments(Provider<? extends Iterable<String>> value)
-
-    void argument(String argument)
-    void arguments(String... arguments)
-    void arguments(Iterable<String> arguments)
-
-
-}
