+++ src/main/groovy/wooga/gradle/snyk/tasks/SnykCheckBase.groovy
@@ -2,6 +2,7 @@ package wooga.gradle.snyk.tasks
 
 import org.gradle.api.tasks.Nested
 import org.gradle.internal.reflect.Instantiator
+import wooga.gradle.snyk.cli.commands.ProjectCommandSpec
 import wooga.gradle.snyk.cli.commands.TestProjectCommandSpec
 import wooga.gradle.snyk.cli.options.CommonOption
 import wooga.gradle.snyk.cli.options.ProjectOption
@@ -11,7 +12,7 @@ import wooga.gradle.snyk.report.SnykReportsImpl
 
 import javax.inject.Inject
 
-abstract class SnykCheckBase extends SnykTask implements TestProjectCommandSpec {
+abstract class SnykCheckBase extends SnykTask implements TestProjectCommandSpec, ProjectCommandSpec {
     @Inject
     protected Instantiator getInstantiator() {
         throw new UnsupportedOperationException()
@@ -31,16 +32,16 @@ abstract class SnykCheckBase extends SnykTask implements TestProjectCommandSpec 
         reports.sarif.outputLocation.convention(project.layout.buildDirectory.file(new File(this.temporaryDir, "report.sarif").absolutePath))
         reports.json.outputLocation.convention(project.layout.buildDirectory.file(new File(this.temporaryDir, "report.json").absolutePath))
 
-        sarifOutputPath.convention(providers.provider({
-            if (reports.sarif.enabled) {
-                return reports.sarif.outputLocation.get()
+        sarifOutputPath.convention(reports.sarif.required.flatMap({
+            if(it) {
+                return reports.sarif.outputLocation
             }
             null
         }))
 
-        jsonOutputPath.convention(providers.provider({
-            if (reports.json.enabled) {
-                return reports.json.outputLocation.get()
+        jsonOutputPath.convention(reports.json.required.flatMap({
+            if(it) {
+                return reports.json.outputLocation
             }
             null
         }))
